Fix regression in characters and costumes

+++ NScumm.Core/Graphics/ClassicCostumeRenderer.cs
@@ -280,8 +280,8 @@ namespace NScumm.Core.Graphics
                         _height = _loaded.CostumeReader.ReadByte();
                         xmoveCur = _xmove + _loaded.CostumeReader.ReadByte() * 8;
                         ymoveCur = _ymove - _loaded.CostumeReader.ReadByte();
-                        _xmove += _loaded.CostumeReader.ReadByte() * 8;
-                        _ymove -= _loaded.CostumeReader.ReadByte();
+                        _xmove += (sbyte)_loaded.CostumeReader.ReadByte() * 8;
+                        _ymove -= (sbyte)_loaded.CostumeReader.ReadByte();
                         _srcptr += 6;
                     }
                     else
